create one base client for node.js and browser
